Extract errors to errors/fs
