+++ packages/hardhat-utils/src/fs.ts
@@ -1,45 +1,16 @@
 import fsPromises from "node:fs/promises";
 import path from "node:path";
 
-import {
-  CustomError,
-  assertHardhatUtilsInvariant,
-} from "./errors/custom-errors.js";
+import { assertHardhatUtilsInvariant } from "./errors/custom-errors.js";
 import { ensureError } from "./errors/catch-utils.js";
-
-// We use this error to encapsulate any other error possibly thrown by node's
-// fs apis, as sometimes their errors don't have stack traces.
-export class FileSystemAccessError extends CustomError {}
-
-export class FileNotFoundError extends CustomError {
-  constructor(filePath: string, cause?: Error) {
-    super(`File ${filePath} not found`, cause);
-  }
-}
-
-export class FileAlreadyExistsError extends CustomError {
-  constructor(filePath: string, cause?: Error) {
-    super(`File ${filePath} already exists`, cause);
-  }
-}
-
-export class InvalidFileFormatError extends CustomError {
-  constructor(filePath: string, cause: Error) {
-    super(`Invalid file format: ${filePath}`, cause);
-  }
-}
-
-export class JsonSerializationError extends CustomError {
-  constructor(filePath: string, cause: Error) {
-    super(`Error serializing JSON file ${filePath}`, cause);
-  }
-}
-
-export class InvalidDirectoryError extends CustomError {
-  constructor(filePath: string, cause: Error) {
-    super(`Invalid directory ${filePath}`, cause);
-  }
-}
+import {
+  FileNotFoundError,
+  FileSystemAccessError,
+  InvalidFileFormatError,
+  JsonSerializationError,
+  FileAlreadyExistsError,
+  InvalidDirectoryError,
+} from "./errors/fs.js";
 
 /**
  * Determines the canonical pathname for a given path, resolving any symbolic
